Pruebas de unidad Forum

+++ forum/grails-app/domain/forum/File.groovy
@@ -4,7 +4,7 @@ class File {
 
     String fileType
     Byte[] content
-    Double size
+    Integer size
 
     static constraints = {
         fileType nullable: false, match: "[a-zA-Z0-9]+'/'[a-zA-Z0-9]+"
